Рефакторинг клиента, id теперь long, как и в остальных сущностях

+++ src/main/java/ustis/fitnesscentre/mapper/ClientMapper.java
@@ -11,7 +11,7 @@ public class ClientMapper implements RowMapper<Client> {
     @Override
     public Client mapRow(ResultSet rs, int rowNum) throws SQLException {
         return new Client(
-                rs.getInt("id"),
+                rs.getLong("id"),
                 rs.getString("phoneNumber"),
                 rs.getString("password"),
                 rs.getString("full_name"),
